requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-telegram-bot==20.3
python-dotenv==1.0.0
fake-useragent==1.2.1
requests==2.31.0